backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
agno
mistralai
sqlalchemy
psycopg[binary]
pgvector
sentence-transformers
chonkie[semantic]
openai
huggingface-hub
transformers
tokenizers
